Minor UI fix and improvements to favorites and comments
This commit starts to address issue #1. Each EMTLPhotoCell now displays
the number of likes and favorites. If the string of names who have
favorites the image will fit on a single line, it displays the names
instead of the number.

This commit fixes is #10. The app now runs full screen and goes beneath
the translucent status bar.